feat: Add voucher system and profile page
Introduce a new voucher system with associated types and data.
Enhance the User type to include a list of acquired vouchers.
Add a dedicated profile page to the application, accessible from the navigation.
Include banner data and facility information in constants for use in UI elements.
Update MenuItem type to support descriptions.